Move BindgenFn into separate module, add BindgenStruct
