+++ cs-bindgen-shared/src/bindgen_struct.rs
@@ -1,0 +1,4 @@
+use serde::*;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct BindgenStruct {}
