+++ cs-bindgen-shared/src/item.rs
@@ -1,0 +1,8 @@
+use crate::{BindgenFn, BindgenStruct};
+use serde::*;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum BindgenItem {
+    Fn(BindgenFn),
+    Struct(BindgenStruct),
+}
